LFS-743: As an API consumer, I can access the human-readable answers in addition to the answer value in the form JSON Move all common methods in the parent class Make the common parent class abstract
