+++ modules/data-entry/src/main/java/ca/sickkids/ccm/lfs/dataentry/internal/serialize/labels/AnswerOptionsLabelProcessor.java
@@ -31,7 +31,6 @@ import javax.json.Json;
 import javax.json.JsonObjectBuilder;
 import javax.json.JsonValue;
 
-import org.apache.sling.api.resource.Resource;
 import org.osgi.service.component.annotations.Component;
 
 import ca.sickkids.ccm.lfs.serialize.spi.ResourceJsonProcessor;
@@ -44,44 +43,6 @@ import ca.sickkids.ccm.lfs.serialize.spi.ResourceJsonProcessor;
 @Component(immediate = true)
 public class AnswerOptionsLabelProcessor extends SimpleAnswerLabelProcessor implements ResourceJsonProcessor
 {
-    @Override
-    public String getName()
-    {
-        return "labels";
-    }
-
-    @Override
-    public int getPriority()
-    {
-        return 75;
-    }
-
-    @Override
-    public boolean isEnabledByDefault(Resource resource)
-    {
-        return true;
-    }
-
-    @Override
-    public boolean canProcess(Resource resource)
-    {
-        return resource.isResourceType("lfs/Form");
-    }
-
-    @Override
-    public JsonValue processChild(final Node node, final Node child, final JsonValue input,
-        final Function<Node, JsonValue> serializeNode)
-    {
-        try {
-            if (child.isNodeType("lfs:TextAnswer")) {
-                return serializeNode.apply(child);
-            }
-        } catch (RepositoryException e) {
-            // Really shouldn't happen
-        }
-        return input;
-    }
-
     @Override
     public void leave(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
     {
