+++ modules/data-entry/src/main/java/ca/sickkids/ccm/lfs/dataentry/internal/serialize/labels/BooleanLabelProcessor.java
@@ -27,7 +27,6 @@ import javax.json.JsonObjectBuilder;
 import javax.json.JsonValue;
 
 import org.apache.commons.lang3.BooleanUtils;
-import org.apache.sling.api.resource.Resource;
 import org.osgi.service.component.annotations.Component;
 
 import ca.sickkids.ccm.lfs.serialize.spi.ResourceJsonProcessor;
@@ -45,44 +44,6 @@ public class BooleanLabelProcessor extends SimpleAnswerLabelProcessor implements
     private static final String NO_LABEL = "noLabel";
 
     private static final String UNKNOWN_LABEL = "unknownLabel";
-
-    @Override
-    public String getName()
-    {
-        return "labels";
-    }
-
-    @Override
-    public int getPriority()
-    {
-        return 75;
-    }
-
-    @Override
-    public boolean isEnabledByDefault(Resource resource)
-    {
-        return true;
-    }
-
-    @Override
-    public boolean canProcess(Resource resource)
-    {
-        return resource.isResourceType("lfs/Form");
-    }
-
-    @Override
-    public JsonValue processChild(final Node node, final Node child, final JsonValue input,
-        final Function<Node, JsonValue> serializeNode)
-    {
-        try {
-            if (child.isNodeType("lfs:BooleanAnswer")) {
-                return serializeNode.apply(child);
-            }
-        } catch (RepositoryException e) {
-            // Really shouldn't happen
-        }
-        return input;
-    }
 
     @Override
     public void leave(Node node, JsonObjectBuilder json, Function<Node, JsonValue> serializeNode)
